+++ src/frontend/apps/impress/src/layouts/PageLayout.tsx
@@ -36,7 +36,7 @@ export function PageLayout({
         $css={css`
           flex-grow: 1;
           &:focus {
-            outline: 3px solid var(--c--theme--colors--primary-600);
+            outline: 3px solid var(--c--globals--colors--primary-600);
             outline-offset: -3px;
           }
           &:focus:not(:focus-visible) {